chore: improve Flex exports
